TEL-14941: Make parsing more permissive to non-conformant SIP
Summary:
In certain cases the Content-Header is mandatory. This parser throws an
error if there is not exactly one Content-Header in such a situation.
We've observed this in some rare instances involving international
carriers. The error results in missing entries in the SIP ladder diagram.

Since the output of this parser operates the diagnostic tool that presents
the SIP ladder diagram,  there is not much to be gained by this kind of
strictness. We can instead treat no Content-Header as Content-
Header = 0. This is probably the correct assumption in these rare cases,
and even if it's not we will have a complete ladder diagram.

While we're at it, we also treat multiple Content-Headers also as
Content-Header = 0.

It would have been nice to add a test to the built-in unit tests. But
I find that the unit tests are broken beyond repair. I've therefore
added a separate file that tests just this case.

Test Plan: Unit test

Reviewers: mghoneim, amigolin

Reviewed By: amigolin

Subscribers: TelPlat, TelCore

Differential Revision: https://phabricator.fspot.us/D17744